% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_query.R
\name{get_query}
\alias{get_query}
\title{Get query from database}
\usage{
get_query(query)
}
\arguments{
\item{query}{string}
}
\value{
data frame
}
\description{
Execute an input SQL query and close the database connection after
}
\examples{
get_query("SELECT TOP (10)
MatrixPlateBC,
DWPlateBC,
RTimestampPM
FROM PlateFlow_Biomek
ORDER BY RTimestampPM DESC")
}
