% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{location_labels}
\alias{location_labels}
\title{Labels for TCDK locations}
\format{
A \code{\link[tibble:tibble]{tibble}} with 2 rows and 3 variables:
\describe{
\item{LocationID}{An integer for joining with database query results}
\item{LocationName}{A human-friendly label (e.g. in plots) for locations}
\item{LocatioNFilename}{Lowercase, UTF-8-friendly string for use in filenames}
}
}
\usage{
location_labels
}
\description{
A dataset for interpreting LocationID values (integers) as human- and
UTF-8 filename-friendly labels.
}
\examples{
d <- get_query(
  "SELECT TOP 10
  MatrixPlateBC,
  LocationID
FROM PlateFlow_Biomek

ORDER BY RTimestampPM DESC"
)

d$LocName <- location_labels$LocationName[d$LocationID]

d

}
\keyword{datasets}
