% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{modes}
\alias{modes}
\alias{get_mode}
\alias{set_mode}
\title{Get tcdkhelper mode}
\usage{
get_mode()

set_mode(mode)
}
\arguments{
\item{mode}{A string giving the mode of operation. When \code{mode} is omitted,
a mode is set automatically. Currently supported values are "RStudio",
"batch" and "compile".}
}
\value{
A string giving the current (or new) tcdkhelper mode.
}
\description{
\code{set_mode()} sets the mode of operation of \code{tcdkhelper}, mainly
affecting relative paths interpreted by \code{\link[=save_kpi]{save_kpi()}}.
}
\details{
Currently, the following modes of operation is implemented in the
\code{save_kpi()} function, and decide where (and if) a plot is saved:

"RStudio" saves to a subfolder of your current working directory called
"KPI-PLOTS" (created if not already existing).

"batch" saves to the folder KPI-PLOTS on a network destination
("S-drive") shared within the TestCenter Danmark organisation.

"compile" does not save the file at all.
This is useful if you have manually opened a graphics device (e.g. by \code{\link[grDevices:pdf]{pdf()}}
and wish to print the plot to given device.

When \code{set_mode()} is called with no \code{mode} argument, a mode is decided as
follows: If \code{save_kpi()} is called within RStudio, the mode "RStudio" is set.
If not, and a graphics device - other than the
"\code{\link[grDevices:dev]{null device}}" - is set, mode is set to "compile".
This mode generates no output, allowing for manually printing plots to
set device, minimizing the risk of duplicate plots in your destination.
Finally, if run outside RStudio with no graphics device, "batch" mode is set.
}
