% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rtimestamp_levels}
\alias{rtimestamp_levels}
\title{Labels for timestamps}
\format{
A \code{\link[tibble:tibble]{tibble}} with 5 rows and 3 variables:
\describe{
\item{Level}{Name of the corresponding RTimestamp column.
Note that this column is not a factor (see details).}
\item{Label}{An easily interpretable string describing what has happened
since previous \code{Level}.}
\item{Hex}{A hex color code (derived from a previously used palette).
Colors of processing steps that are similar in nature (e.g. RNA extraction
with liquid handling robots as opposed to rt-PCR and human interpretation
of these results), and any potential future timestamp should most likely
adhere to these "rules".}
}
}
\usage{
rtimestamp_levels
}
\description{
A dataset for arranging timestamps from PlateFlow tables such as PlateFlow_Biomek
with rows ordered by order of step in processing.
This allows factoring a "long-format" dataframe (timestamp names and values given
in two columns) so that downstream plotting maintains correct processing order.

Also gives (suggested) labels, describing what has been done since
previous timestamp.
}
\details{
It is (potentially) important to note that the column \code{Level} is
\emph{not} a factor.
This is due to how joins work in \code{\link[dplyr:mutate-joins]{dplyr}}; joining a
(e.g. \code{chr}) vector with a \code{factor} coerces the matching column to a vector
despite the join method used.

To circumvent this, a target column (existing or new) can be made a factor by
using the \code{Level} column for levels.
}
\examples{
timestamps <- tibble::tibble(
  Timestamp = c(
    "RTimestampBKi7",
    "RTimestampPCR",
    "RTimestampPCRVV",
    "RTimestampPCRDV",
    "RTimestampPCRRes"
  ),
  Time = c(
    "1999",
    "Sunday Bloody Sunday",
    "9 to 5",
    "Yesterday",
    "Right Here Right Now"
  )
)

timestamps$Timestamp <- factor(timestamps$Timestamp,
                               levels = rtimestamp_levels$Level,
                               labels = rtimestamp_levels$Label)

timestamps

}
\keyword{datasets}
