% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_kpi.R
\name{save_kpi}
\alias{save_kpi}
\title{Save KPI plot}
\usage{
save_kpi(
  path,
  format = "screen",
  mode = get_mode(),
  margin = 0.17,
  plot = ggplot2::last_plot(),
  dpi = 300,
  force_path = F,
  portrait = F,
  debug = F
)
}
\arguments{
\item{path}{File path}

\item{format}{A string giving output format ("a3", "a4", or "screen")}

\item{mode}{Mode of operation. See \link{modes} for more.}

\item{margin}{Margin of plot (in inches). Default of 0.17 matches most printers.}

\item{plot}{ggplot2 plot}

\item{dpi}{A number giving DPI}

\item{force_path}{bool}

\item{portrait}{bool}

\item{debug}{bool}
}
\description{
Save a plot to a number of friendly formats
}
